remove more unused services, move scene service to projects folder
