+++ packages/server-core/src/projects/scene/scene-helper.ts
