+++ packages/server-core/src/projects/scene/scene-parser.ts
