+++ packages/server-core/src/projects/scene/scene-schema-parse.test.ts
