+++ packages/server-core/src/projects/scene/scene.class.ts
@@ -5,7 +5,7 @@ import fs from 'fs'
 import path from 'path'
 import appRootPath from 'app-root-path'
 import { cleanString } from '../../util/cleanString'
-import { uploadLocalProjectToProvider } from '../../projects/project/project.class'
+import { uploadLocalProjectToProvider } from '../project/project.class'
 import { isDev } from '@xrengine/common/src/utils/isDev'
 import defaultSceneSeed from '@xrengine/projects/default-project/empty.scene.json'
 import { useStorageProvider } from '../../media/storageprovider/storageprovider'
