+++ packages/server-core/src/projects/scene/scene.docs.ts
