+++ packages/server-core/src/projects/scene/scene.hooks.ts
