+++ packages/server-core/src/projects/scene/scene.service.ts
