+++ packages/server-core/src/projects/services.ts
@@ -1,4 +1,5 @@
 import Project from './project/project.service'
 import GithubApp from './githubapp/githubapp.service'
+import Scene from './scene/scene.service'
 
-export default [Project, GithubApp]
+export default [GithubApp, Project, Scene]
