--- packages/server-core/src/world/meta/meta.docs.ts
@@ -1,17 +1,0 @@
-/**
- * An object for swagger documentation configiration
- *
- * @author Kevin KIMENYI
- */
-export default {
-  definitions: {
-    meta: {
-      type: 'object',
-      properties: {}
-    },
-    meta_list: {
-      type: 'array',
-      items: { $ref: '#/definitions/meta' }
-    }
-  }
-}
