--- packages/server-core/src/world/meta/meta.hooks.ts
@@ -1,33 +1,0 @@
-import { disallow } from 'feathers-hooks-common'
-
-export default {
-  before: {
-    all: [],
-    find: [],
-    get: [disallow()],
-    create: [disallow()],
-    update: [disallow()],
-    patch: [disallow()],
-    remove: [disallow()]
-  },
-
-  after: {
-    all: [],
-    find: [],
-    get: [],
-    create: [],
-    update: [],
-    patch: [],
-    remove: []
-  },
-
-  error: {
-    all: [],
-    find: [],
-    get: [],
-    create: [],
-    update: [],
-    patch: [],
-    remove: []
-  }
-} as any
