--- packages/server-core/src/world/resolve-media/resolve-media.docs.ts
@@ -1,17 +1,0 @@
-/**
- * An object for swagger documentation configiration
- *
- * @author Kevin KIMENYI
- */
-export default {
-  definitions: {
-    'resolve-media': {
-      type: 'object',
-      properties: {}
-    },
-    'resolve-media_list': {
-      type: 'array',
-      items: { $ref: '#/definitions/resolve-media' }
-    }
-  }
-}
