--- packages/server-core/src/world/resolve-media/resolve-media.hooks.ts
@@ -1,33 +1,0 @@
-import { disallow } from 'feathers-hooks-common'
-
-export default {
-  before: {
-    all: [],
-    find: [disallow()],
-    get: [disallow()],
-    create: [],
-    update: [disallow()],
-    patch: [disallow()],
-    remove: [disallow()]
-  },
-
-  after: {
-    all: [],
-    find: [],
-    get: [],
-    create: [],
-    update: [],
-    patch: [],
-    remove: []
-  },
-
-  error: {
-    all: [],
-    find: [],
-    get: [],
-    create: [],
-    update: [],
-    patch: [],
-    remove: []
-  }
-} as any
