--- packages/server-core/src/world/seeder-config.ts
@@ -1,5 +1,0 @@
-import { ServicesSeedConfig } from '@xrengine/common/src/interfaces/ServicesSeedConfig'
-
-export const worldSeeds: Array<ServicesSeedConfig> = []
-
-export default worldSeeds
