--- packages/server-core/src/world/services.ts
@@ -1,5 +1,0 @@
-import Scene from './scene/scene.service'
-import Meta from './meta/meta.service'
-import ResolveMedia from './resolve-media/resolve-media.service'
-
-export default [ResolveMedia, Scene, Meta]
